Refactor: 스킬선택 Select 내부에서 직접 스킬 api요청하기
