+++ webapp/src/components/TechStackSelectInput/TechStackSelectInput.stories.jsx
@@ -1,17 +1,21 @@
-import { skillStack } from 'constant';
+import handlers from 'mocks/handlers';
 import React from 'react';
 import { skillStackParser } from 'service/skillStack.parser';
-import TechStackSelectInput from './index';
+import { skillStack } from 'constant';
+import TechStackSelectInputView from './TechStackSelectInput.view';
 
 export default {
-  title: 'Component/TechStackSelectInput',
-  component: TechStackSelectInput,
+  title: 'Component/TechStackSelectInputView',
+  component: TechStackSelectInputView,
   args: {},
+  parameters: {
+    msw: handlers,
+  },
   layout: 'fullscreen',
 };
 
 function Template(args) {
-  return <TechStackSelectInput {...args} />;
+  return <TechStackSelectInputView {...args} />;
 }
 
 export const Default = Template.bind({});
@@ -26,10 +30,11 @@ Default.args = {
         'https://user-images.githubusercontent.com/71386219/186051220-a77fa08e-b501-4baa-af3c-47ae602d25e1.png',
     },
   ],
-  name: 'TechStackSelectInput',
-  label: 'TechStackSelectInput',
+  name: 'techSkills',
+  label: '기술스택선택하기',
   techSkillOptions: skillStackParser(skillStack),
   placeholder: 'selectedTechSkills',
+  skillStack,
 };
 
 export const WithError = Template.bind({});
@@ -41,4 +46,5 @@ WithError.args = {
   placeholder: 'Error',
   isError: true,
   helperText: '에러 입니다.',
+  skillStack,
 };
